requirements.txt
Django==1.11.6
pytz==2017.3
djangorestframework==3.7.3
qrcode==5.3
reportlab==3.4.0
olefile==0.44
Pillow==4.3.0
pytz==2017.3
six==1.11.0
psycopg2==2.7.3.2
django_cors_headers==2.1.0
drf-nested-routers
gunicorn
